requirements.txt
streamlit
crewai
crewai-tools
python-dotenv
google-generativeai
langchain-community
langchain-google-genai
pyyaml